Base Roles Support
Two options were available, first provide a way for the SP to filter and
return a set of roles from the global pool back to the IdP, or as this is
doing using domain scoped roles.  This is somewhat easier as it requires
Keystone interaction rather than Sentinel code.  The SP creates a domain
role for users which then uses the inference API to reference policy
roles.  It also provides a nice mechanism to abstract away where an SP
requires multiple roles for things like base membership, object stores and
heat etc which can be inferred from a single role.  Creating, updating and
deletion of roles is denied as without SP intervention it is pretty much
useless.