Refactor Tests To Use Fixtures
There is a lot of duplication to do with creating and destroying test resources
which is a problem solved by using fixtures in test cases.  This is especially
so for tests that require many resources, testing role assignment for example
requires users, roles, projects and a grant.  This can all be succinctly wrapped
up in an aggregate fixture.