Create REST Controller Base Class
Somewhere to place common functionality.  In this instance formatting and whitelisting
is pushed up into the base.  This allows us to extract the collection and resource
class variables defined in concrete controllers to implicitly define the white list
operation without the backtrace hack, it still may be explicitly overridden for custom
actions which return different entities that the main bulk of operations.