Client Code Refactor
References have been updated to be compute and identity, not nova or keystone.
Access to clients is now done via a base REST controller property which lazily
creates the clients on a per-request basis.  We could cache across requests
but I'm unsure as to what happens in the event of a token expiry, I'm guessing
a 403... As a result the concrete controllers are rid of a lot of duplicated
client instantiations.