Improve Exception Handling
Certain OpenStack clients e.g nova, expect JSON error messages to have been
formatted by webob's exception handler.  Pecan does its own formatting of
webob exceptions which gets in the way.  To remedy the situation if we undo
this by deleting the pecan generated body in an error hook then the exception
will be called to generate the body and we receive the correct formatting.