Check Query Parameters
Controllers which may accept query parameters have not protection at present.  This
means a client may expect certain behaviour and get something back with a good
status code.  This adds in a decorator which whitelists supported query parameters,
if we recieve any not supported it will raise a HTTP 400 describing which arguments
are not supported.