Fix Neutron Client Model
As Neutron doesn't return resources but rather the raw output it imposed some
nasty conditions upon this code which make client interoperability pretty much
impossible.  As such I've wrapped the client in a class which intercepts all
calls to get attributes.  If it's a function call this is also wrapped so that
API calls returning dicts can be marshalled into resources so they behave in
the same way as keystone and nova resources and I can revert the hacks.  It's
been implmented as a class decorator so can be easily reused if other clients
misbehave.