Meters API
Implements the meters API used internally to do billing.  Due to Fog using
the body for queries and the official clients using the http query string
it seemed simpler to let WSME handle the paramters much like how they are
managed natively in ceilometer.

Identity user can now have a password set, however this is discouraged as it
may go out of sync with the federated identity provider.  This allows us to
create a federated user client, and subsequently a server fixture to test
metrics injected into ceilometer.