Improve Fitration
Extends the Scope.filter API to accept an explicit list of projects
to filter against and an object attribute to compare against.  Default
behaviour is still to select all projects in the domain and implicitly
look for project_id/tenant_id attributes.  This allows code to be shared
with glance (which uses owner), cinder (which is optionall project scoped
and uses os-vol-tenant-attr:tenant_id) and nova (which is optionally
project scoped)